85 feat: invisible form field
An invisible form field is useful to carry invisible, computable data
from a higher input stage, such as a yad list widget piped into the
form -- to the form output stage.

The form does not display the invisible form field, and the user cannot
interact with it. However, form button "@"commands receive and can
change its value, and the value is output together with all other
form field values.